Create sub-controller for general osci-render settings
